break out G3tSmurf tables and add max_ctime search
